requirements.txt
click==8.1.7
python-slugify==8.0.1
text-unidecode==1.3
